Split SyncEngine
